requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.7
django-push-notifications==1.2.0
django-toolbelt==0.0.1
django-uuidfield==0.5.0
gunicorn==19.3.0
pbr==0.10.8
psycopg2==2.5.2
pyOpenSSL==0.14
requests==2.6.0
six==1.9.0
South==1.0.2
static3==0.5.1
stevedore==1.3.0
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
